ok, fix bug in build

+++ paper1/misc/build.py
@@ -92,10 +92,6 @@ def build_figures():
             ]
         )
         error = bool(len(meta.get(name, {}).get("stderr", "")))
-        if error:
-            # Fail silently
-            print("ERROR running {}:".format(name))
-            print(meta.get(name, {}).get("stderr", ""))
         if stale or missing or error:
             old_figures = set(
                 glob.glob("figures/*.pdf") + glob.glob("figures/*.png")
@@ -110,6 +106,9 @@ def build_figures():
             except subprocess.CalledProcessError as e:
                 stdout = ""
                 stderr = e.output.decode("utf-8")
+                # Fail silently
+                print("ERROR running {}:".format(name))
+                print(stderr)
 
             new_figures = set(
                 glob.glob("figures/*.pdf") + glob.glob("figures/*.png")
@@ -200,7 +199,8 @@ def build():
 def clean(remove_data=False):
     # Remove figure output
     for file in glob.glob("figures/*.pdf") + glob.glob("figures/*.png"):
-        os.remove(file)
+        if not "greetings.png" in file:  # HACK: keep this one
+            os.remove(file)
     # Remove test output
     for file in glob.glob("tests/*.tex"):
         os.remove(file)
